remove index file
