--- apps/explorer/src/ui/RingChart/index.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export { RingChart, type RingChartProps } from './RingChart';
-export { RingChartLegend } from './RingChartLegend';
